python-auto-generate (NEW DESIGN)
Somehow we lost the idea of automatically generating the CUDA wrappers.
That made the code extremely difficult to develop and to debug, since
there are about 100 CUDA functions.

A tag, old-no-auto-generate, has been created for that obsolete design.

With this new design, we will again auto-generate:
  cuda_plugin.h cudaproxy.icu cudawrappers.icpp
in the python-auto-generate subdirectory.  We can then modify the main
Makefile to copy those files into the plugin directory and integrate
them appropriately.